Merge from:      branches/cotillion/merge/add_npc/      branches/cotillion/merge/query_interactive/
Adding (to rooms):
  add_npc
  add_object 

These will add a monster/object to a room and handle then handle the 
reset call also!

To this I had to add the query_reset_active funtion to /std/object also,  
as reset has to be started for this to work.

Changing: 
query_ip_* to query_interactive where appropriate




git-svn-id: https://genesis.tekno.chalmers.se:443/svn/mudlib/trunk@136 a2e4861b-4ae5-0310-b681-bd0aab12f737