This is a set of balance fixes.
The mana formula is changed to increase the mana recovery
rate to offset the loss of mass mana herb consumtion somewhat
as per the AoB board note. WIS is used as the base for mana
recovery instead of int, int decides how much mana you have.


Herbs can only be consumed at a rate of one per 60 seconds.
This should probably not be considered a 'permanent' solution
as it's still possible to create too powerful herbs. 




git-svn-id: https://genesis.tekno.chalmers.se:443/svn/mudlib/trunk@186 a2e4861b-4ae5-0310-b681-bd0aab12f737