This fixes a bug report from 1999 where players would turn to attack themselves after killing a foe.
The routine for new target selections was moved to a new function to 
remove code duplication.

A problem where cb_query_attack would return objects not in the room
should also be fixed, this is the likely cause of a number of odd
looking runtimes.

cb_heal is also removed since it has not worked for many years.



git-svn-id: https://genesis.tekno.chalmers.se:443/svn/mudlib/trunk@292 a2e4861b-4ae5-0310-b681-bd0aab12f737