Fix 'say to' to prefer targeting livings over items
'say to human It is sunny' parsing results in an incorrect match
when there is also an object with human as adjective in the room.

'human it' would then be consumed by %i despite the primary match
being on 'human' with 'It ..' as the text.

This is likely a driver issue but fixing the parse_command API will
cause alot of issues.