Fix the default addition of plurals to prevent double pluralization
When a name is set with set/add_name a plural name is also added
unless the noplural flag is set. This would often result in
incorrect plural names when the object was a pair or set.

A pair of bracers would create the plural name 'bracerses'
To preven that a new language macro LANG_IS_PLURAL has been
added with some naive rules to detect plurals.